requirement.txt
paho-mqtt
Adafruit-DHT
opencv-python
RPi.GPIO
